initial booth combinations
